+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Checkbox condition now checks if number is valid
+- Checkbox now uses `checked` also
+
 ## [3.5.7] - 2019-04-22
 
 ### Fixed